% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_portfolio.R
\name{new_portfolio_site}
\alias{new_portfolio_site}
\title{Creates a New Data Science Portfolio Website}
\usage{
new_portfolio_site(
  dir = ".",
  force = NA,
  install_hugo = TRUE,
  sample = FALSE,
  format = "yaml",
  theme = c("raditian", "kross"),
  theme_example = TRUE,
  netlify = TRUE,
  ...
)
}
\arguments{
\item{dir}{The directory of the new site.}

\item{force}{Whether to create the site in a directory even if it is not
empty. By default, \code{force = TRUE} when the directory only contains
hidden, RStudio project (\file{*.Rproj}), \file{LICENSE}, and/or
\file{README} files.}

\item{install_hugo}{Whether to install Hugo automatically if it is not found.}

\item{sample}{Whether to add sample content. Hugo creates an empty site by
default, but this function adds sample content by default.}

\item{format}{The format of the configuration file, e.g., \code{'yaml'} or
\code{'toml'} (the value \code{TRUE} will be treated as \code{'yaml'}, and
\code{FALSE} means \code{'toml'}). Note that the frontmatter of the new (R)
Markdown file created by \code{new_content()} always uses YAML instead of
TOML or JSON.}

\item{theme}{The theme can be defined in several ways.

\strong{Preset themes}
\itemize{
\item \code{raditian} (default): Custom version of the free hugo raditian theme
\item \code{kross}: Custom version of the free hugo kross theme
}

\strong{A Hugo theme on Github}
\itemize{
\item A character string of the form user/repo
\item A GIT branch or tag name after @, i.e. theme can be of the form user/repo@branch).
\item A full URL to the zip file or tarball of the theme.
}

\strong{No Theme}

If \code{theme = NA}, no themes will be installed, and you have to manually install a theme.}

\item{theme_example}{Whether to copy the example in the \file{exampleSite}
directory if it exists in the theme. Not all themes provide example sites.}

\item{netlify}{Whether to create a Netlify config file \file{netlify.toml}.}

\item{...}{Additional arguments can be passed to \code{\link[blogdown:hugo_cmd]{blogdown::new_site()}}.}
}
\description{
\itemize{
\item Generates the default hugo template skeleton for a Portfolio-Based Data Scientist Website
\item Provides 2 starter portfolio templates: \code{raditian} and \code{kross}.
}
}
