% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{build_site}
\alias{serve_site}
\alias{stop_server}
\alias{use_github}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{blogdown}{\code{\link[blogdown]{build_site}}, \code{\link[blogdown]{serve_site}}, \code{\link[blogdown:serve_site]{stop_server}}}

  \item{usethis}{\code{\link[usethis]{use_github}}}
}}

